ghg-sustainability-app/requirements.txt
# Core Framework
streamlit==1.35.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Authentication
bcrypt==4.1.1

# Data Processing
pandas==2.1.3
openpyxl==3.1.2

# PDF Generation
reportlab==4.0.7

# Email
email-validator==2.1.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Environment Variables
python-dotenv==1.0.0

# Type Hints
typing-extensions==4.8.0

# Utilities
pathlib2==2.3.7.post1
